refactor: pull out ser module for note
